styling
additional styling to the results page, there still needs to be work done, but this is a big push forward towards the final product.

The only thing that needs to be changed now is the alignment of the cards on the page, they need to be center, and the standardization of the card size.